# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Writeo API Documentation
  version: 1.0.0
  description: |
    # Writeo API Documentation

    Comprehensive API documentation for the Writeo essay assessment platform.

    ## Overview

    Writeo provides two main APIs:

    1. **Text API** (Cloudflare Workers) - Public-facing API for creating questions, answers, and submissions
    2. **Essay Scorer API** (Modal/FastAPI) - Internal ML service for essay scoring

    ## Architecture

    - **Edge API**: Cloudflare Workers for low-latency global access
    - **Storage**: R2 bucket for questions/answers/submissions, KV namespace for results
    - **ML Service**: Modal-hosted FastAPI service with transformer models for essay scoring
    - **Models**: Multiple scoring models available (engessay, distilbert, fallback)

    ## Authentication

    All endpoints (except `/health`, `/docs`, `/openapi.json`) require API key authentication via the `Authorization` header:
    ```
    Authorization: Token <api_key>
    ```

    ## Rate Limits

    Rate limits are applied per IP address:

    - **Submissions**: 5 requests per minute (expensive operations)
    - **Results (GET)**: 60 requests per minute (read-only)
    - **Questions**: 30 requests per minute (data writes)
    - **Other endpoints**: 30 requests per minute

    Contact support for higher limits.

  contact:
    name: Writeo API Support
    url: https://github.com/your-org/writeo
  license:
    name: Apache-2.0
...
servers:
...
paths:
...
  /text/questions/{question_id}:
    put:
      tags: [Questions]
      summary: Create or update a question
      description: |
        Creates a new question or updates an existing one if the content is identical.
        Questions are stored in R2 storage.
      operationId: createQuestion
      security:
        - ApiKeyAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          description: Unique identifier for the question (UUID format)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: The question text
                  example: "Describe your weekend. What did you do?"
      responses:
        "201":
          description: Question created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Question created"
        "204":
          description: Question already exists with identical content
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid question_id format"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Question already exists with different content"
        "500":
          description: Internal server error
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API key authentication using the format: `Token <api_key>`

        Example: `Authorization: Token your-api-key-here`

        All endpoints except `/health`, `/docs`, and `/openapi.json` require this authentication.